requirements/prod.txt
SQLAlchemy==2.0.35
sqlalchemy-utils==0.41.2
aiogram==3.13.1
asyncpg==0.29.0
pydantic==2.9.2
pyyaml==6.0.2
alembic==1.13.2
